Created helper class for category sorting
