+++ Plugin/Block/Toolbar.php
@@ -1,6 +1,6 @@
 <?php
 /**
- * Copyright (c) 2017, Nosto Solutions Ltd
+ * Copyright (c) 2019, Nosto Solutions Ltd
  * All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification,
@@ -29,16 +29,16 @@
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  *
  * @author Nosto Solutions Ltd <contact@nosto.com>
- * @copyright 2017 Nosto Solutions Ltd
+ * @copyright 2019 Nosto Solutions Ltd
  * @license http://opensource.org/licenses/BSD-3-Clause BSD 3-Clause
  *
  */
 
 namespace Nosto\Tagging\Plugin\Block;
 
-use Nosto\Tagging\Plugin\Catalog\Model\Config;
 use Nosto\Tagging\Helper\Data as NostoHelperData;
 use Nosto\Tagging\Helper\Account as NostoHelperAccount;
+use Nosto\Tagging\Helper\CategorySorting as NostoHelperSorting;
 use Magento\Backend\Block\Template\Context;
 use Magento\Store\Model\StoreManagerInterface;
 use Magento\Catalog\Block\Product\ProductList\Toolbar as MagentoToolbar;
@@ -135,9 +135,10 @@ class Toolbar extends Template
         $this->_collection = $collection;
         $store = $this->storeManager->getStore();
         $currentOrder = $subject->getCurrentOrder();
-        if (($currentOrder === Config::NOSTO_PERSONALIZED_KEY
-            || $currentOrder === Config::NOSTO_TOPLIST_KEY)
+        if (($currentOrder === NostoHelperSorting::NOSTO_PERSONALIZED_KEY
+            || $currentOrder === NostoHelperSorting::NOSTO_TOPLIST_KEY)
             && $this->nostoHelperAccount->nostoInstalledAndEnabled($store)
+            && $this->nostoHelperData->isCategorySortingEnabled($store)
             && $this->nostoHelperData->isCategorySortingEnabled($store)
         ) {
             $limit = (int)$subject->getLimit();
@@ -170,6 +171,9 @@ class Toolbar extends Template
     private function getSortedIds(Store $store, $type)
     {
         $nostoAccount = $this->nostoHelperAccount->findAccount($store);
+        if ($nostoAccount === null) {
+            return [];
+        }
         $category = $this->registry->registry('current_category');
         $category = $this->categoryBuilder->build($category, $store);
         $nostoCustomer = $this->cookieManager->getCookie(NostoCustomer::COOKIE_NAME);
